Fix some gcc wshadow warnings
